+++ packages/substream/sink/utils/ipfs.ts
@@ -1,24 +1,8 @@
-import { Duration, Effect, Either, Schedule } from 'effect';
+import { Duration, Effect, Schedule } from 'effect';
 import type { TimeoutException } from 'effect/Cause';
 
 import { IPFS_GATEWAY } from '../constants/constants';
-import { Spaces } from '../db';
-import { SpaceWithPluginAddressNotFoundError } from '../errors';
-import {
-  type ContentProposal,
-  type EditorshipProposal,
-  type MembershipProposal,
-  type ProposalCreated,
-  type SubspaceProposal,
-  ZodContentProposal,
-  ZodMembershipProposal,
-  ZodProposalMetadata,
-  ZodSubspaceProposal,
-} from '../events/proposals-created/parser';
 import { type UriData } from '../zod';
-import { isValidAction } from './actions';
-import { getChecksumAddress } from './get-checksum-address';
-import { slog } from './slog';
 
 class UnableToParseBase64Error extends Error {
   _tag: 'UnableToParseBase64Error' = 'UnableToParseBase64Error';
@@ -98,201 +82,5 @@ export function getFetchIpfsContentEffect(
 
     // We only support IPFS URIs or base64 encoded content with the above format
     return null;
-  });
-}
-
-/**
- * We don't know the content type of the proposal until we fetch the IPFS content and parse it.
- *
- * 1. Verify that the space exists for the plugin address
- * 2. Fetch the IPFS content
- * 3. Parse the IPFS content to get the "type"
- * 4. Return an object matching the type and the expected contents.
- *
- * Later on we map this to the database schema and write the proposal to the database.
- */
-export function getProposalFromMetadata(
-  proposal: ProposalCreated
-): Effect.Effect<
-  ContentProposal | SubspaceProposal | MembershipProposal | EditorshipProposal | null,
-  SpaceWithPluginAddressNotFoundError
-> {
-  return Effect.gen(function* (unwrap) {
-    // The proposal can come from either the voting plugin or the membership plugin
-    // depending on which type of proposal is being processed
-    const maybeSpaceIdForVotingPlugin = yield* unwrap(
-      Effect.promise(() => Spaces.findForVotingPlugin(proposal.pluginAddress))
-    );
-    const maybeSpaceIdForMembershipPlugin = yield* unwrap(
-      Effect.promise(() => Spaces.findForMembershipPlugin(proposal.pluginAddress))
-    );
-
-    if (!maybeSpaceIdForVotingPlugin && !maybeSpaceIdForMembershipPlugin) {
-      slog({
-        message: `Matching space in Proposal not found for plugin address ${proposal.pluginAddress}`,
-        requestId: '-1',
-      });
-
-      return null;
-    }
-
-    slog({
-      message: `Fetching IPFS content for proposal
-        proposalId:    ${proposal.proposalId}
-        pluginAddress: ${proposal.pluginAddress}
-        creator:       ${proposal.creator}
-        metadataUri:   ${proposal.metadataUri}
-        startTime:     ${proposal.startTime}
-        endTime:       ${proposal.endTime}`,
-      requestId: '-1',
-    });
-
-    const fetchIpfsContentEffect = getFetchIpfsContentEffect(proposal.metadataUri);
-    const maybeIpfsContent = yield* unwrap(Effect.either(fetchIpfsContentEffect));
-
-    if (Either.isLeft(maybeIpfsContent)) {
-      const error = maybeIpfsContent.left;
-
-      switch (error._tag) {
-        case 'UnableToParseBase64Error':
-          console.error(`Unable to parse base64 string ${proposal.metadataUri}`, error);
-          break;
-        case 'FailedFetchingIpfsContentError':
-          console.error(`Failed fetching IPFS content from uri ${proposal.metadataUri}`, error);
-          break;
-        case 'UnableToParseJsonError':
-          console.error(`Unable to parse JSON when reading content from uri ${proposal.metadataUri}`, error);
-          break;
-        case 'TimeoutException':
-          console.error(`Timed out when fetching IPFS content for uri ${proposal.metadataUri}`, error);
-          break;
-        default:
-          console.error(`Unknown error when fetching IPFS content for uri ${proposal.metadataUri}`, error);
-          break;
-      }
-
-      return null;
-    }
-
-    const ipfsContent = maybeIpfsContent.right;
-
-    if (!ipfsContent) {
-      return null;
-    }
-
-    const validIpfsMetadata = ZodProposalMetadata.safeParse(ipfsContent);
-
-    if (!validIpfsMetadata.success) {
-      // @TODO: Effectify error handling
-      console.error('Failed to parse IPFS metadata', validIpfsMetadata.error);
-      return null;
-    }
-
-    switch (validIpfsMetadata.data.type) {
-      case 'CONTENT': {
-        const parsedContent = ZodContentProposal.safeParse(ipfsContent);
-
-        // Subspace proposals are only emitted by the voting plugin
-        if (!parsedContent.success || !maybeSpaceIdForVotingPlugin) {
-          return null;
-        }
-
-        const mappedProposal: ContentProposal = {
-          ...proposal,
-          type: validIpfsMetadata.data.type,
-          name: validIpfsMetadata.data.name ?? null,
-          proposalId: parsedContent.data.proposalId,
-          onchainProposalId: proposal.proposalId,
-          actions: parsedContent.data.actions.filter(isValidAction),
-          creator: getChecksumAddress(proposal.creator),
-          space: getChecksumAddress(maybeSpaceIdForVotingPlugin),
-          // json: JSON.stringify(ipfsContent),
-          // uri: proposal.metadataUri,
-        };
-
-        return mappedProposal;
-      }
-
-      case 'ADD_SUBSPACE':
-      case 'REMOVE_SUBSPACE': {
-        // @TODO: ipfs content type is not correct for non-content-type proposals
-        const parsedSubspace = ZodSubspaceProposal.safeParse(ipfsContent);
-
-        // Subspace proposals are only emitted by the voting plugin
-        if (!parsedSubspace.success || !maybeSpaceIdForVotingPlugin) {
-          return null;
-        }
-
-        const mappedProposal: SubspaceProposal = {
-          ...proposal,
-          type: validIpfsMetadata.data.type,
-          name: validIpfsMetadata.data.name ?? null,
-          proposalId: parsedSubspace.data.proposalId,
-          onchainProposalId: proposal.proposalId,
-          subspace: getChecksumAddress(parsedSubspace.data.subspace),
-          creator: getChecksumAddress(proposal.creator),
-          space: getChecksumAddress(maybeSpaceIdForVotingPlugin),
-          // json: JSON.stringify(ipfsContent),
-          // uri: proposal.metadataUri,
-        };
-
-        return mappedProposal;
-      }
-
-      case 'ADD_EDITOR':
-      case 'REMOVE_EDITOR':
-        const parsedMembership = ZodMembershipProposal.safeParse(ipfsContent);
-
-        if (!parsedMembership.success) {
-          return null;
-        }
-
-        // If both of these are null then we already early exit out of this function, so it's safe to cast
-        // to the correct type here.
-        const spaceAddress = (maybeSpaceIdForMembershipPlugin ?? maybeSpaceIdForVotingPlugin) as `0x${string}`;
-
-        const mappedProposal: EditorshipProposal = {
-          ...proposal,
-          type: validIpfsMetadata.data.type,
-          name: validIpfsMetadata.data.name ?? null,
-          proposalId: parsedMembership.data.proposalId,
-          onchainProposalId: proposal.proposalId,
-          userAddress: getChecksumAddress(parsedMembership.data.userAddress),
-          creator: getChecksumAddress(proposal.creator),
-          space: getChecksumAddress(spaceAddress),
-          // json: JSON.stringify(ipfsContent),
-          // uri: proposal.metadataUri,
-        };
-
-        return mappedProposal;
-
-      case 'ADD_MEMBER':
-      case 'REMOVE_MEMBER': {
-        const parsedMembership = ZodMembershipProposal.safeParse(ipfsContent);
-
-        if (!parsedMembership.success) {
-          return null;
-        }
-
-        // If both of these are null then we already early exit out of this function, so it's safe to cast
-        // to the correct type here.
-        const spaceAddress = (maybeSpaceIdForMembershipPlugin ?? maybeSpaceIdForVotingPlugin) as `0x${string}`;
-
-        const mappedProposal: MembershipProposal = {
-          ...proposal,
-          type: validIpfsMetadata.data.type,
-          name: validIpfsMetadata.data.name ?? null,
-          proposalId: parsedMembership.data.proposalId,
-          onchainProposalId: proposal.proposalId,
-          userAddress: getChecksumAddress(parsedMembership.data.userAddress),
-          creator: getChecksumAddress(proposal.creator),
-          space: getChecksumAddress(spaceAddress),
-          // json: JSON.stringify(ipfsContent),
-          // uri: proposal.metadataUri,
-        };
-
-        return mappedProposal;
-      }
-    }
   });
 }
